+++ index.html
@@ -28,7 +28,7 @@
     <link rel="stylesheet" href="css/style.css" />
 
     <!-- js -->
-    <script src="js/main.js"></script>
+    <script src="./js/main.js" type="module"></script>
   </head>
 
   <body>
