+++ js/createExhibitorNode.js
@@ -1,0 +1,6 @@
+export const createExhibitorNode = (exhibitor) => {
+  const span = document.createElement("span");
+  const name = document.createTextNode(exhibitor.name);
+  span.appendChild(name);
+  return span;
+};
