+++ js/insertExhibitor.js
@@ -1,0 +1,7 @@
+export const insertExhibitor = (exhibitors) => {
+  exhibitors.forEach((exhibitor) => {
+    document
+      .getElementById(exhibitor.labo + "-labo")
+      .appendChild(createExhibitorNode(exhibitor));
+  });
+};
