+++ js/insertLecturer.js
@@ -1,0 +1,7 @@
+export const insertLecturer = (lecturers) => {
+  lecturers.forEach((lecturer) => {
+    document
+      .getElementById("lecturer-profile-ul")
+      .appendChild(createLecturerNode(lecturer));
+  });
+};
