+++ js/main.js
@@ -1,3 +1,8 @@
+import { insertExhibitor } from "./insertExhibitor";
+import { createExhibitorNode } from "./createExhibitorNode";
+import { insertLecturer } from "./insertLecturer";
+import { createLecturerNode } from "./createLecturerNode";
+
 // 学生データ挿入
 fetch("./json/exhibitor.json")
   .then((response) => response.json())
@@ -39,65 +44,65 @@ window.onload = () => {
   });
 };
 
-const insertExhibitor = (exhibitors) => {
-  exhibitors.forEach((exhibitor) => {
-    document
-      .getElementById(exhibitor.labo + "-labo")
-      .appendChild(createExhibitorNode(exhibitor));
-  });
-};
+// const insertExhibitor = (exhibitors) => {
+//   exhibitors.forEach((exhibitor) => {
+//     document
+//       .getElementById(exhibitor.labo + "-labo")
+//       .appendChild(createExhibitorNode(exhibitor));
+//   });
+// };
 
-const createExhibitorNode = (exhibitor) => {
-  const span = document.createElement("span");
-  const name = document.createTextNode(exhibitor.name);
-  span.appendChild(name);
-  return span;
-};
+// const createExhibitorNode = (exhibitor) => {
+//   const span = document.createElement("span");
+//   const name = document.createTextNode(exhibitor.name);
+//   span.appendChild(name);
+//   return span;
+// };
 
-const insertLecturer = (lecturers) => {
-  lecturers.forEach((lecturer) => {
-    document
-      .getElementById("lecturer-profile-ul")
-      .appendChild(createLecturerNode(lecturer));
-  });
-};
+// const insertLecturer = (lecturers) => {
+//   lecturers.forEach((lecturer) => {
+//     document
+//       .getElementById("lecturer-profile-ul")
+//       .appendChild(createLecturerNode(lecturer));
+//   });
+// };
 
-const createLecturerNode = (lecturer) => {
-  const img = document.createElement("img");
-  img.setAttribute("class", "lecturer-profile-img");
-  img.setAttribute("src", "./img/lecturers/" + lecturer.img_name);
+// const createLecturerNode = (lecturer) => {
+//   const img = document.createElement("img");
+//   img.setAttribute("class", "lecturer-profile-img");
+//   img.setAttribute("src", "./img/lecturers/" + lecturer.img_name);
 
-  const roman = document.createTextNode(lecturer.roman);
-  const romanSpan = document.createElement("span");
-  romanSpan.setAttribute("class", "lecturer-profile-text");
-  romanSpan.appendChild(roman);
+//   const roman = document.createTextNode(lecturer.roman);
+//   const romanSpan = document.createElement("span");
+//   romanSpan.setAttribute("class", "lecturer-profile-text");
+//   romanSpan.appendChild(roman);
 
-  const name = document.createTextNode(lecturer.name);
-  const nameSpan = document.createElement("span");
-  nameSpan.setAttribute("class", "lecturer-profile-name");
-  nameSpan.appendChild(name);
+//   const name = document.createTextNode(lecturer.name);
+//   const nameSpan = document.createElement("span");
+//   nameSpan.setAttribute("class", "lecturer-profile-name");
+//   nameSpan.appendChild(name);
 
-  const belongs = document.createTextNode(lecturer.belongs);
-  const belongsSpan = document.createElement("span");
-  belongsSpan.setAttribute("class", "lecturer-profile-text");
-  belongsSpan.appendChild(belongs);
+//   const belongs = document.createTextNode(lecturer.belongs);
+//   const belongsSpan = document.createElement("span");
+//   belongsSpan.setAttribute("class", "lecturer-profile-text");
+//   belongsSpan.appendChild(belongs);
 
-  const div = document.createElement("div");
-  div.appendChild(img);
-  div.appendChild(romanSpan);
-  div.appendChild(nameSpan);
-  div.appendChild(belongsSpan);
+//   const div = document.createElement("div");
+//   div.appendChild(img);
+//   div.appendChild(romanSpan);
+//   div.appendChild(nameSpan);
+//   div.appendChild(belongsSpan);
 
-  const career = document.createTextNode(
-    lecturer.position + "  " + lecturer.career
-  );
-  const p = document.createElement("p");
-  p.appendChild(career);
+//   const career = document.createTextNode(
+//     lecturer.position + "  " + lecturer.career
+//   );
+//   const p = document.createElement("p");
+//   p.appendChild(career);
 
-  const li = document.createElement("li");
-  li.setAttribute("class", "lecturer-profile-li");
-  li.appendChild(div);
-  li.appendChild(p);
+//   const li = document.createElement("li");
+//   li.setAttribute("class", "lecturer-profile-li");
+//   li.appendChild(div);
+//   li.appendChild(p);
 
-  return li;
-};
+//   return li;
+// };
